Make setup easier
There are only two system libraries that this project needs: ffmpeg
and python3-pip. The rest are just python libraries listed in the
requirements.txt in `pip freeze` format. Previously, `ui.sh install`
was written to simplify installation, but I forgot to add system
libraries to it. This commit adds system libraries along with the
version of the python libraries, so a new user can install this
project way easier.